Bug-Fix: access operations with loops in the CFG (#1558)
* feat-fix(cfg): acess without faulty loops

* lint-fix(cfg): handle linter errors
